using math library
Finding the square root of a number is the inverse operation of squaring that number. Remember, the square of a number is that number times itself. The perfect squares are the squares of the whole numbers. The square root of a number, n, written below is the number that gives n when multiplied by itself.