Create nCr & nPr.
In NCR and NPR, C stands for Combinations, and P stands for permutations. Now for combinations, it is the number of ways you can pick r objects out of n. In permutations, it is the number of ways you can pick r objects out of n and then arrange them. Therefore, NCR= selecting r from n.