Create sort the strings.
Sort an array of strings based on the frequency of good words in them. Unbounded Fractional Knapsack. Rearrange the characters of the string such that no two adjacent characters are consecutive English alphabets. Sort an array of strings according to string lengths using Map. Remove elements to make array sorted.