Create  add two matrix.
Sum of two matrix: Sum of two Matrix is a matrix obtained by adding the corresponding elements of the two given matrix. But to perform the Sum of two Matrix the number of rows and columns of two matrix must be equal. Thus the Sum matrix will have same number of rows and columns as the two given matrix